Query serializer can be reused
